requirements.txt
mysql-connector-python
python-dotenv
flask_sqlalchemy
flask